added ./tacspeak/test_model.py; removed ./test_model.py
